refactor: clean up more code and remove Point
Replaced float with double

+++ RailwayCo/Assets/Scripts/GameLogic/Train/TrainAttribute.cs
@@ -1,19 +1,19 @@
 ﻿public class TrainAttribute
 {
     private Attribute<int> trainCapacity;
-    private Attribute<float> trainFuel;
-    private Attribute<float> trainDurability;
-    private Attribute<float> trainSpeed;
+    private Attribute<double> trainFuel;
+    private Attribute<double> trainDurability;
+    private Attribute<double> trainSpeed;
 
     public Attribute<int> TrainCapacity { get => trainCapacity; private set => trainCapacity = value; }
-    public Attribute<float> TrainFuel { get => trainFuel; private set => trainFuel = value; }
-    public Attribute<float> TrainDurability { get => trainDurability; private set => trainDurability = value; }
-    public Attribute<float> TrainSpeed { get => trainSpeed; private set => trainSpeed = value; }
+    public Attribute<double> TrainFuel { get => trainFuel; private set => trainFuel = value; }
+    public Attribute<double> TrainDurability { get => trainDurability; private set => trainDurability = value; }
+    public Attribute<double> TrainSpeed { get => trainSpeed; private set => trainSpeed = value; }
 
     public void UpgradeCapacityLimit(int capacityLimit) => TrainCapacity.UpperLimit += capacityLimit;
-    public void UpgradeFuelRate(float fuelRate) => TrainFuel.Rate += fuelRate;
-    public void UpgradeFuelLimit(float fuelLimit) => TrainFuel.UpperLimit += fuelLimit;
-    public void UpgradeDurabilityRate(float durabilityRate) => TrainDurability.Rate += durabilityRate;
-    public void UpgradeDurabilityLimit(float durabilityLimit) => TrainDurability.UpperLimit += durabilityLimit;
-    public void UpgradeSpeedLimit(float speedLimit) => TrainSpeed.UpperLimit += speedLimit;
+    public void UpgradeFuelRate(double fuelRate) => TrainFuel.Rate += fuelRate;
+    public void UpgradeFuelLimit(double fuelLimit) => TrainFuel.UpperLimit += fuelLimit;
+    public void UpgradeDurabilityRate(double durabilityRate) => TrainDurability.Rate += durabilityRate;
+    public void UpgradeDurabilityLimit(double durabilityLimit) => TrainDurability.UpperLimit += durabilityLimit;
+    public void UpgradeSpeedLimit(double speedLimit) => TrainSpeed.UpperLimit += speedLimit;
 }
